Refactor the interface IProps of ReactMultiEmail
Divide Interface code to simplify ReactMultiEmail.tsx
