+++ packages/react-multi-email/IProps.tsx
@@ -1,0 +1,18 @@
+import * as React from 'react'
+
+interface IProps {
+    emails?: string[];
+    onChange?: (emails: string[]) => void;
+    noClass?: boolean;
+    validateEmail?: (email: string) => boolean;
+    style?: object;
+    getLabel: (
+      email: string,
+      index: number,
+      removeEmail: (index: number) => void,
+    ) => void;
+    className?: string;
+    placeholder?: string | React.ReactNode;
+  }
+
+  export default IProps;
